revert int->long change as it breaks on 64bit. Actually only VM_CallInterpreted hurts as VM_CallCompiled is only used on 32bit. Use same arg types for consistency nevertheless.

+++ code/qcommon/vm_local.h
@@ -168,10 +168,10 @@ extern	vm_t	*currentVM;
 extern	int		vm_debugLevel;
 
 void VM_Compile( vm_t *vm, vmHeader_t *header );
-int	VM_CallCompiled( vm_t *vm, long *args );
+int	VM_CallCompiled( vm_t *vm, int *args );
 
 void VM_PrepareInterpreter( vm_t *vm, vmHeader_t *header );
-int	VM_CallInterpreted( vm_t *vm, long *args );
+int	VM_CallInterpreted( vm_t *vm, int *args );
 
 vmSymbol_t *VM_ValueToFunctionSymbol( vm_t *vm, int value );
 int VM_SymbolToValue( vm_t *vm, const char *symbol );
