+++ code/qcommon/vm_none.c
@@ -1,6 +1,6 @@
 #include "vm_local.h"
 
-int VM_CallCompiled( vm_t *vm, long *args ) {
+int VM_CallCompiled( vm_t *vm, int *args ) {
 	exit(99);
 	return 0;
 }
